Refactored 'app_config' module.
It now have public and private implmentation details, with only the public part being published.

Using `resourceGroup()' as a scope when calling a module resulted in a dependeny being added to the deployment, rather than the deployment scope being set.

This does not happen when using a symbolic reference to a resource group.

All other modules are unchanged and currently treated public, though their implementations may need to be reviewed.

Updated build process to only publish 'public' modules.